requirements.txt
torch>=1.6.0
torchvision>=0.7.0
matplotlib
sklearn
hydra-core
omegaconf
pyyaml
tqdm
